refactor: move pkg/jpath and pkg/native into pkg/jsonnet
Both packages are jsonnet related so it makes sense to organize them as
subpackages of the `pkg/jsonnet` package
